mv content of messages.py to output_utils.py, with minor refactoring and no caching
